[testing] <-- [372 ingredientservice refractor]

+++ TasteITServer/src/main/java/pl/jakubkonkol/tasteitserver/service/IngredientService.java
@@ -67,15 +67,27 @@ public class IngredientService {
         });
     }
     public void deleteAll() {
+        if (ingredientRepository.count() == 0) {
+            throw new IllegalStateException("No ingredients to delete.");
+        }
         ingredientRepository.deleteAll();
     }
 
     public void deleteById(String ingredientId) {
+        if (ingredientId == null || ingredientId.isEmpty()) {
+            throw new IllegalArgumentException("Ingredient ID cannot be null or empty.");
+        }
+        if (!ingredientRepository.existsById(ingredientId)) {
+            throw new IllegalStateException("Ingredient with ID " + ingredientId + " does not exist.");
+        }
         ingredientRepository.deleteById(ingredientId);
     }
+
     public List<IngredientDto> getAll() {
         List<Ingredient> ingredients = ingredientRepository.findAll();
-        return ingredients.stream().map(this::convertToDto).toList();
+        return ingredients.stream()
+                .map(this::convertToDto)
+                .toList();
     }
 
     //na razie nie potrzebna
